Postdump of Bootstrap Posts: Ready for Mallet(LDA)
The boostrap table was simply a selection of all posts with the tag
`bootstrap`.

	INSERT INTO bootstrap select * from posts where posts.Tags like
	'%<bootstrap>%';

The Python script loads all the rows from the resulting table and strips
html tags and special characters. The left-over body is then dumped to
the data folder in a text file name as `{post-id}.txt`.

An example of topics resulting from Mallet is given in the topic_keys
file.

Signed-off-by: Martin Jorn Rogalla <martin@martinrogalla.com>